backend/requirements.txt
fastapi
uvicorn[standard]
openai>=1.0.0
langchain
langchain-community
chromadb
Pillow
python-dotenv
pydantic
python-multipart
httpx